vid-app-common/src/main/java/org/onap/vid/job/command/ResourceInProgressStatusCommand.java
/*-
 * ============LICENSE_START=======================================================
 * VID
 * ================================================================================
 * Copyright (C) 2017 - 2019 AT&T Intellectual Property. All rights reserved.
 * ================================================================================
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *      http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * ============LICENSE_END=========================================================
 */

package org.onap.vid.job.command;

import org.onap.vid.job.Job;
import org.onap.vid.job.NextCommand;
import org.onap.vid.job.impl.JobSharedData;
import org.springframework.beans.factory.config.ConfigurableBeanFactory;
import org.springframework.context.annotation.Scope;
import org.springframework.stereotype.Component;

@Component
@Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
public class ResourceInProgressStatusCommand extends BaseInProgressStatusCommand {

    public ResourceInProgressStatusCommand() {
    }

    ResourceInProgressStatusCommand(JobSharedData sharedData, String requestId, String instanceId) {
        init(sharedData, requestId, instanceId);
    }

    @Override
    protected ExpiryChecker getExpiryChecker() {
        return x->false;
    }

    @Override
    protected NextCommand processJobStatus(Job.JobStatus jobStatus) {
        return new NextCommand(jobStatus, this);
    }


}
